// app/types.ts
export type BreadCrumbItemType = {
    label?: string;
    url?: string;
}

export type ProductType =  {
    id: number;
    title: string;
    description: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    brand: string;
    category: string;
    thumbnail: string;
    images: string[];
  }

export type cartItemType = {
    id:number,
    quantity:number
}